fix: Rework on dialog add
