--- app/src/main/java/fr/upec/e2ee/ui/gallery/GalleryViewModel.java
@@ -1,21 +1,0 @@
-package fr.upec.e2ee.ui.gallery;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class GalleryViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public GalleryViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("contacts");
-
-
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
