+++ app/src/main/res/layout/contacts_dialog.xml
@@ -19,7 +19,7 @@
         android:layout_marginTop="16dp"
         android:layout_marginRight="4dp"
         android:layout_marginBottom="4dp"
-        android:hint="@string/Username"
+        android:hint="@string/name"
         android:inputType="textEmailAddress" />
 
     <EditText
