+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -13,9 +13,9 @@
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="fr.upec.e2ee.ui.gallery.GalleryFragment"
+        android:name="fr.upec.e2ee.ui.directory.DirectoryFragment"
         android:label="@string/menu_gallery"
-        tools:layout="@layout/fragment_gallery" />
+        tools:layout="@layout/fragment_directory" />
 
     <fragment
         android:id="@+id/nav_slideshow"
